Extract native sidebar components
